Polish
Removed In from parameters to be more inlign with mosr code bases even
though I prefer being explicit.
Improved alire configuration and gpr setup and moved any code that does
not sit well with a light runtime into Original_Source. In the case of
secondary stack use then these should be easily fixed in time with
performance gains. In the case of recursion then a bit more thought will
be required but it should be fairly straight forward, again as required.